55: extract the remaining symbols, which are related to auth, into the newly-created `frontend/src/features/auth/` directory
