+++ frontend/src/features/auth/Login.tsx
@@ -4,14 +4,13 @@ import { ThunkDispatch } from "redux-thunk";
 import { Redirect, Link } from "react-router-dom";
 import { v4 as uuidv4 } from "uuid";
 
+import { IState, selectHasValidToken } from "../../store";
 import {
-  IState,
   issueJWSToken,
   ActionIssueJWSToken,
   fetchProfile,
   ActionFetchProfile,
-  selectHasValidToken,
-} from "../../store";
+} from "./authSlice";
 import { IActionAlertsCreate, alertsCreate } from "../alerts/alertsSlice";
 
 export const Login = () => {
