+++ frontend/src/store.test.ts
@@ -1,4 +1,5 @@
-import { IAlert } from "./types";
+import { IAlert, RequestStatus, IProfile } from "./types";
+import { INITIAL_STATE_AUTH } from "./constants";
 
 // 1
 import {
@@ -6,14 +7,6 @@ import {
   INITIAL_STATE,
   selectAlertsIds,
   selectAlertsEntities,
-  RequestStatus,
-  createUserPending,
-  createUserRejected,
-  createUserFulfilled,
-  IActionCreateUserPending,
-  ActionTypesCreateUser,
-  IActionCreateUserRejected,
-  IActionCreateUserFulfilled,
 } from "./store";
 
 // 2
@@ -23,42 +16,8 @@ import thunkMiddleware, { ThunkDispatch } from "redux-thunk";
 import { AnyAction } from "redux";
 import configureMockStore, { MockStoreEnhanced } from "redux-mock-store";
 
-import { createUser } from "./store";
-
 // 3
-import { IStateAuth, initialStateAuth, authReducer } from "./store";
-
-import {
-  ActionTypesIssueJWSToken,
-  issueJWSTokenPending,
-  IActionIssueJWSTokenPending,
-  issueJWSTokenRejected,
-  IActionIssueJWSTokenRejected,
-  issueJWSTokenFulfilled,
-  IActionIssueJWSTokenFulfilled,
-  issueJWSToken,
-} from "./store";
-
-import {
-  VOCAB_TREASURY_APP_TOKEN,
-  ACTION_TYPE_AUTH_CLEAR_SLICE,
-  IActionAuthClearSlice,
-  authClearSlice,
-  logOut,
-} from "./store";
-
-import {
-  ActionTypesFetchProfile,
-  fetchProfilePending,
-  IActionFetchProfilePending,
-  fetchProfileRejected,
-  IActionFetchProfileRejected,
-  fetchProfileFulfilled,
-  IActionFetchProfileFulfilled,
-  IProfile,
-} from "./store";
 import { profileMock } from "./dataMocks";
-import { fetchProfile } from "./store";
 
 import {
   selectAuthRequestStatus,
@@ -138,14 +97,6 @@ import {
   IActionEditExampleRejected,
   IActionEditExampleFulfilled,
   editExample,
-  requestPasswordResetPending,
-  requestPasswordResetRejected,
-  requestPasswordResetFulfilled,
-  ActionTypesRequestPasswordReset,
-  IActionRequestPasswordResetPending,
-  IActionRequestPasswordResetRejected,
-  IActionRequestPasswordResetFulfilled,
-  requestPasswordReset,
 } from "./store";
 
 describe("selector functions", () => {
@@ -163,7 +114,7 @@ describe("selector functions", () => {
         },
       },
       auth: {
-        ...initialStateAuth,
+        ...INITIAL_STATE_AUTH,
         requestStatus: RequestStatus.SUCCEEDED,
         loggedInUserProfile: {
           id: 17,
@@ -301,131 +252,6 @@ describe("selector functions", () => {
 });
 
 describe("action creators", () => {
-  test("createUserPending", () => {
-    const action = createUserPending();
-
-    expect(action).toEqual({
-      type: "auth/createUser/pending",
-    });
-  });
-
-  test("createUserRejected", () => {
-    const action = createUserRejected("auth-createUser-rejected");
-
-    expect(action).toEqual({
-      type: "auth/createUser/rejected",
-      error: "auth-createUser-rejected",
-    });
-  });
-
-  test("createUserFulfilled", () => {
-    const action = createUserFulfilled();
-
-    expect(action).toEqual({
-      type: "auth/createUser/fulfilled",
-    });
-  });
-
-  test("issueJWSTokenPending", () => {
-    const action = issueJWSTokenPending();
-
-    expect(action).toEqual({
-      type: "auth/issueJWSToken/pending",
-    });
-  });
-
-  test("issueJWSTokenRejected", () => {
-    const action = issueJWSTokenRejected("auth-issueJWSToken-rejected");
-
-    expect(action).toEqual({
-      type: "auth/issueJWSToken/rejected",
-      error: "auth-issueJWSToken-rejected",
-    });
-  });
-
-  test("issueJWSTokenFulfilled", () => {
-    const action = issueJWSTokenFulfilled("token-issued-by-the-backend");
-
-    expect(action).toEqual({
-      type: "auth/issueJWSToken/fulfilled",
-      payload: {
-        token: "token-issued-by-the-backend",
-      },
-    });
-  });
-
-  test("fetchProfilePending", () => {
-    const action = fetchProfilePending();
-
-    expect(action).toEqual({
-      type: "auth/fetchProfile/pending",
-    });
-  });
-
-  test("fetchProfileRejected", () => {
-    const action = fetchProfileRejected("auth-fetchProfile-rejected");
-
-    expect(action).toEqual({
-      type: "auth/fetchProfile/rejected",
-      error: "auth-fetchProfile-rejected",
-    });
-  });
-
-  test("fetchProfileFulfilled", () => {
-    const profile: IProfile = {
-      id: 17,
-      username: "mocked-jd",
-      email: "mocked-john.doe@protonmail.com",
-    };
-    const action = fetchProfileFulfilled(profile);
-
-    expect(action).toEqual({
-      type: "auth/fetchProfile/fulfilled",
-      payload: {
-        profile: {
-          id: 17,
-          username: "mocked-jd",
-          email: "mocked-john.doe@protonmail.com",
-        },
-      },
-    });
-  });
-
-  test("requestPasswordResetPending", () => {
-    const action = requestPasswordResetPending();
-
-    expect(action).toEqual({
-      type: "auth/requestPasswordReset/pending",
-    });
-  });
-
-  test("requestPasswordResetRejected", () => {
-    const action = requestPasswordResetRejected(
-      "auth-requestPasswordReset-rejected"
-    );
-
-    expect(action).toEqual({
-      type: "auth/requestPasswordReset/rejected",
-      error: "auth-requestPasswordReset-rejected",
-    });
-  });
-
-  test("requestPasswordResetFulfilled", () => {
-    const action = requestPasswordResetFulfilled();
-
-    expect(action).toEqual({
-      type: "auth/requestPasswordReset/fulfilled",
-    });
-  });
-
-  test("authClearSlice", () => {
-    const action = authClearSlice();
-
-    expect(action).toEqual({
-      type: "auth/clearSlice",
-    });
-  });
-
   test("fetchExamplesPending", () => {
     const action = fetchExamplesPending();
 
@@ -688,299 +514,6 @@ describe("action creators", () => {
 });
 
 describe("slice reducers", () => {
-  describe("authReducer", () => {
-    test("auth/createUser/pending", () => {
-      const initState: IStateAuth = {
-        ...initialStateAuth,
-        requestStatus: RequestStatus.FAILED,
-        requestError: "auth-createUser-rejected",
-      };
-      const action: IActionCreateUserPending = {
-        type: ActionTypesCreateUser.PENDING,
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        requestStatus: RequestStatus.LOADING,
-        requestError: null,
-        token: null,
-        hasValidToken: null,
-        loggedInUserProfile: null,
-      });
-    });
-
-    test("auth/createUser/rejected", () => {
-      const initState: IStateAuth = {
-        ...initialStateAuth,
-        requestStatus: RequestStatus.LOADING,
-        requestError: null,
-      };
-      const action: IActionCreateUserRejected = {
-        type: ActionTypesCreateUser.REJECTED,
-        error: "auth-createUser-rejected",
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        requestStatus: RequestStatus.FAILED,
-        requestError: "auth-createUser-rejected",
-        token: null,
-        hasValidToken: null,
-        loggedInUserProfile: null,
-      });
-    });
-
-    test("auth/createUser/fulfilled", () => {
-      const initState: IStateAuth = {
-        ...initialStateAuth,
-        requestStatus: RequestStatus.LOADING,
-        requestError: null,
-      };
-      const action: IActionCreateUserFulfilled = {
-        type: ActionTypesCreateUser.FULFILLED,
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        requestStatus: RequestStatus.SUCCEEDED,
-        requestError: null,
-        token: null,
-        hasValidToken: null,
-        loggedInUserProfile: null,
-      });
-    });
-
-    test("auth/issueJWSToken/pending", () => {
-      const initState: IStateAuth = {
-        ...initialStateAuth,
-        requestStatus: RequestStatus.FAILED,
-        requestError: "auth-issueJWSToken-rejected",
-      };
-      const action: IActionIssueJWSTokenPending = {
-        type: ActionTypesIssueJWSToken.PENDING,
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        requestStatus: RequestStatus.LOADING,
-        requestError: null,
-        token: null,
-        hasValidToken: null,
-        loggedInUserProfile: null,
-      });
-    });
-
-    test("auth/issueJWSToken/rejected", () => {
-      const initState: IStateAuth = {
-        ...initialStateAuth,
-        requestStatus: RequestStatus.LOADING,
-        requestError: null,
-      };
-      const action: IActionIssueJWSTokenRejected = {
-        type: ActionTypesIssueJWSToken.REJECTED,
-        error: "auth-issueJWSToken-rejected",
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        requestStatus: RequestStatus.FAILED,
-        requestError: "auth-issueJWSToken-rejected",
-        token: null,
-        hasValidToken: false,
-        loggedInUserProfile: null,
-      });
-    });
-
-    test("auth/issueJWSToken/fulfilled", () => {
-      const initState: IStateAuth = {
-        ...initialStateAuth,
-        requestStatus: RequestStatus.LOADING,
-        requestError: null,
-      };
-      const action: IActionIssueJWSTokenFulfilled = {
-        type: ActionTypesIssueJWSToken.FULFILLED,
-        payload: {
-          token: "token-issued-by-the-backend",
-        },
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        requestStatus: RequestStatus.SUCCEEDED,
-        requestError: null,
-        token: "token-issued-by-the-backend",
-        hasValidToken: true,
-        loggedInUserProfile: null,
-      });
-    });
-
-    test("auth/fetchProfile/pending", () => {
-      const initState: IStateAuth = {
-        ...initialStateAuth,
-        requestStatus: RequestStatus.FAILED,
-        requestError: "auth-fetchProfile-rejected",
-        token: "token-that-was-loaded-from-localStorage-but-is-no-longer-valid",
-      };
-      const action: IActionFetchProfilePending = {
-        type: ActionTypesFetchProfile.PENDING,
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        requestStatus: RequestStatus.LOADING,
-        requestError: null,
-        token: "token-that-was-loaded-from-localStorage-but-is-no-longer-valid",
-        hasValidToken: null,
-        loggedInUserProfile: null,
-      });
-    });
-
-    test("auth/fetchProfile/rejected", () => {
-      const initState: IStateAuth = {
-        ...initialStateAuth,
-        requestStatus: RequestStatus.LOADING,
-        requestError: null,
-        token: "token-that-was-loaded-from-localStorage-but-is-no-longer-valid",
-      };
-      const action: IActionFetchProfileRejected = {
-        type: ActionTypesFetchProfile.REJECTED,
-        error: "auth-fetchProfile-rejected",
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        requestStatus: RequestStatus.FAILED,
-        requestError: "auth-fetchProfile-rejected",
-        token: "token-that-was-loaded-from-localStorage-but-is-no-longer-valid",
-        hasValidToken: false,
-        loggedInUserProfile: null,
-      });
-    });
-
-    test("auth/fetchProfile/fulfilled", () => {
-      const initState: IStateAuth = {
-        ...initialStateAuth,
-        requestStatus: RequestStatus.LOADING,
-        requestError: null,
-        token: "token-that-was-loaded-from-localStorate-and-is-still-valid",
-      };
-      const action: IActionFetchProfileFulfilled = {
-        type: ActionTypesFetchProfile.FULFILLED,
-        payload: {
-          profile: {
-            id: 17,
-            username: "mocked-jd",
-            email: "mocked-john.doe@protonmail.com",
-          },
-        },
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        requestStatus: RequestStatus.SUCCEEDED,
-        requestError: null,
-        token: "token-that-was-loaded-from-localStorate-and-is-still-valid",
-        hasValidToken: true,
-        loggedInUserProfile: {
-          id: 17,
-          username: "mocked-jd",
-          email: "mocked-john.doe@protonmail.com",
-        },
-      });
-    });
-
-    test("auth/requestPasswordReset/pending", () => {
-      const initState: IStateAuth = {
-        ...initialStateAuth,
-        requestStatus: RequestStatus.FAILED,
-        requestError: "auth-requestPasswordReset-rejected",
-      };
-      const action: IActionRequestPasswordResetPending = {
-        type: ActionTypesRequestPasswordReset.PENDING,
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        ...initState,
-        requestStatus: RequestStatus.LOADING,
-        requestError: null,
-      });
-    });
-
-    test("auth/requestPasswordReset/rejected", () => {
-      const initState: IStateAuth = {
-        ...initialStateAuth,
-        requestStatus: RequestStatus.LOADING,
-      };
-      const action: IActionRequestPasswordResetRejected = {
-        type: ActionTypesRequestPasswordReset.REJECTED,
-        error: "auth-requestPasswordReset-rejected",
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        ...initState,
-        requestStatus: RequestStatus.FAILED,
-        requestError: "auth-requestPasswordReset-rejected",
-      });
-    });
-
-    test("auth/requestPasswordReset/fulfilled", () => {
-      const initState: IStateAuth = {
-        ...initialStateAuth,
-        requestStatus: RequestStatus.LOADING,
-      };
-      const action: IActionRequestPasswordResetFulfilled = {
-        type: ActionTypesRequestPasswordReset.FULFILLED,
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        ...initState,
-        requestStatus: RequestStatus.SUCCEEDED,
-      });
-    });
-
-    test("auth/clearSlice", () => {
-      const initState: IStateAuth = {
-        requestStatus: RequestStatus.SUCCEEDED,
-        requestError: null,
-        token: "token-issued-by-the-backend",
-        hasValidToken: true,
-        loggedInUserProfile: {
-          id: 17,
-          username: "mocked-jd",
-          email: "mocked-john.doe@protonmail.com",
-        },
-      };
-      const action: IActionAuthClearSlice = {
-        type: ACTION_TYPE_AUTH_CLEAR_SLICE,
-      };
-
-      const newState: IStateAuth = authReducer(initState, action);
-
-      expect(newState).toEqual({
-        requestStatus: RequestStatus.SUCCEEDED,
-        requestError: null,
-        token: null,
-        hasValidToken: false,
-        loggedInUserProfile: null,
-      });
-    });
-  });
-
   describe("examplesReducer", () => {
     test("examples/fetchExamples/pending", () => {
       const initState: IStateExamples = {
@@ -1723,220 +1256,6 @@ describe(
     });
 
     test(
-      "createUser(username, ...)" +
-        " + the HTTP request issued by that thunk-action is mocked to succeed",
-      async () => {
-        const createUserPromise = storeMock.dispatch(
-          createUser(
-            "mocked-username",
-            "mocked-email@protonmail.com",
-            "mocked-password"
-          )
-        );
-
-        await expect(createUserPromise).resolves.toEqual(undefined);
-        expect(storeMock.getActions()).toEqual([
-          { type: "auth/createUser/pending" },
-          { type: "auth/createUser/fulfilled" },
-        ]);
-      }
-    );
-
-    test(
-      "createUser(username, ...)" +
-        " + the HTTP request issued by that thunk-action is mocked to fail",
-      async () => {
-        quasiServer.use(
-          rest.post("/api/users", (req, res, ctx) => {
-            return res(
-              ctx.status(400),
-              ctx.json({
-                error: "[mocked] Bad Request",
-                message: "[mocked] The provided email is already taken.",
-              })
-            );
-          })
-        );
-
-        const createUserPromise = storeMock.dispatch(
-          createUser(
-            "mocked-username",
-            "mocked-email@protonmail.com",
-            "mocked-password"
-          )
-        );
-
-        await expect(createUserPromise).rejects.toEqual(
-          "[mocked] The provided email is already taken."
-        );
-        expect(storeMock.getActions()).toEqual([
-          {
-            type: "auth/createUser/pending",
-          },
-          {
-            type: "auth/createUser/rejected",
-            error: "[mocked] The provided email is already taken.",
-          },
-        ]);
-      }
-    );
-
-    test(
-      "issueJWSToken(email, password)" +
-        " + the HTTP request issued by that thunk-action is mocked to succeed",
-      async () => {
-        const issueJWSTokenPromise = storeMock.dispatch(
-          issueJWSToken("mocked-email@protonmail.com", "mocked-password")
-        );
-
-        await expect(issueJWSTokenPromise).resolves.toEqual(undefined);
-        expect(storeMock.getActions()).toEqual([
-          {
-            type: "auth/issueJWSToken/pending",
-          },
-          {
-            type: "auth/issueJWSToken/fulfilled",
-            payload: {
-              token: "mocked-token",
-            },
-          },
-        ]);
-      }
-    );
-
-    test(
-      "issueJWSToken(email, password)" +
-        " + the HTTP request issued by that thunk-action is mocked to fail",
-      async () => {
-        quasiServer.use(
-          rest.post("/api/tokens", (req, res, ctx) => {
-            return res(
-              ctx.status(401),
-              ctx.json({
-                error: "[mocked] Unauthorized",
-                message: "[mocked] Incorrect email and/or password.",
-              })
-            );
-          })
-        );
-
-        const issueJWSTokenPromise = storeMock.dispatch(
-          issueJWSToken("mocked-email@protonmail.com", "mocked-password")
-        );
-
-        await expect(issueJWSTokenPromise).rejects.toEqual(
-          "[mocked] Incorrect email and/or password."
-        );
-        expect(storeMock.getActions()).toEqual([
-          {
-            type: "auth/issueJWSToken/pending",
-          },
-          {
-            type: "auth/issueJWSToken/rejected",
-            error: "[mocked] Incorrect email and/or password.",
-          },
-        ]);
-      }
-    );
-
-    test(
-      "fetchProfile()" +
-        " + the HTTP request issued by that thunk-action is mocked to succeed",
-      async () => {
-        const fetchProfilePromise = storeMock.dispatch(fetchProfile());
-
-        await expect(fetchProfilePromise).resolves.toEqual(undefined);
-        expect(storeMock.getActions()).toEqual([
-          {
-            type: "auth/fetchProfile/pending",
-          },
-          {
-            type: "auth/fetchProfile/fulfilled",
-            payload: {
-              profile: profileMock,
-            },
-          },
-        ]);
-      }
-    );
-
-    test(
-      "fetchProfile()" +
-        " + the HTTP request issued by that thunk-action is mocked to fail",
-      async () => {
-        quasiServer.use(
-          rest.get("/api/user-profile", (req, res, ctx) => {
-            return res(
-              ctx.status(401),
-              ctx.json({
-                error: "[mocked] Unauthorized",
-                message: "[mocked] Expired access token.",
-              })
-            );
-          })
-        );
-
-        const fetchProfilePromise = storeMock.dispatch(fetchProfile());
-
-        await expect(fetchProfilePromise).rejects.toEqual(
-          "[mocked] Expired access token."
-        );
-        expect(storeMock.getActions()).toEqual([
-          {
-            type: "auth/fetchProfile/pending",
-          },
-          {
-            type: "auth/fetchProfile/rejected",
-            error: "[mocked] Expired access token.",
-          },
-        ]);
-      }
-    );
-
-    test(
-      "logOut()" +
-        " [Note: (a) not an async thunk-action," +
-        " and (b) does not dependent on mocking of any HTTP requests]",
-      () => {
-        /* Arrange. */
-        localStorage.setItem(
-          VOCAB_TREASURY_APP_TOKEN,
-          "token-issued-by-the-backend"
-        );
-
-        /* Act. */
-        storeMock.dispatch(logOut("LOGOUT SUCCESSFUL"));
-
-        /* Assert. */
-        const dispatchedActions = storeMock.getActions();
-
-        expect(dispatchedActions.length).toEqual(3);
-
-        expect(dispatchedActions[0]).toEqual({
-          type: "auth/clearSlice",
-        });
-
-        expect(dispatchedActions[1]).toEqual({
-          type: "examples/clearSlice",
-        });
-
-        expect({
-          type: dispatchedActions[2].type,
-          payload: {
-            message: dispatchedActions[2].payload.message,
-          },
-        }).toEqual({
-          type: "alerts/create",
-          payload: {
-            message: "LOGOUT SUCCESSFUL",
-          },
-        });
-
-        expect(localStorage.getItem(VOCAB_TREASURY_APP_TOKEN)).toEqual(null);
-      }
-    );
-
-    test(
       "fetchExamples()" +
         " + the HTTP request issued by that thunk-action is mocked to succeed",
       async () => {
@@ -2223,61 +1542,6 @@ describe(
           {
             type: "examples/editExample/rejected",
             error: "[mocked] Expired access token.",
-          },
-        ]);
-      }
-    );
-
-    test(
-      "requestPasswordReset(email)" +
-        " + the HTTP request issued by that thunk-action is mocked to succeed",
-      async () => {
-        const requestPasswordResetPromise = storeMock.dispatch(
-          requestPasswordReset("mocked-email@protonmail.com")
-        );
-
-        await expect(requestPasswordResetPromise).resolves.toEqual(undefined);
-        expect(storeMock.getActions()).toEqual([
-          {
-            type: "auth/requestPasswordReset/pending",
-          },
-          {
-            type: "auth/requestPasswordReset/fulfilled",
-          },
-        ]);
-      }
-    );
-
-    test(
-      "requestPasswordReset(email)" +
-        " + the HTTP request issued by that thunk-action is mocked to fail",
-      async () => {
-        quasiServer.use(
-          rest.post("/api/request-password-reset", (req, res, ctx) => {
-            return res(
-              ctx.status(400),
-              ctx.json({
-                error: "[mocked] Bad Request",
-                message: "[mocked] This is a bad request.",
-              })
-            );
-          })
-        );
-
-        const requestPasswordResetPromise = storeMock.dispatch(
-          requestPasswordReset("mocked-email@protonmail.com")
-        );
-
-        await expect(requestPasswordResetPromise).rejects.toEqual(
-          "[mocked] This is a bad request."
-        );
-        expect(storeMock.getActions()).toEqual([
-          {
-            type: "auth/requestPasswordReset/pending",
-          },
-          {
-            type: "auth/requestPasswordReset/rejected",
-            error: "[mocked] This is a bad request.",
           },
         ]);
       }
